fix : page.tsx , layout.tsx만 남도록 수정했습니다"
