+++ src/app/(home)/layout.tsx
@@ -1,0 +1,16 @@
+import TopLogoBar from "@/components/ui/TopLogoBar";
+
+interface HomeLayoutProps {
+  children: React.ReactNode;
+}
+
+const HomeLayout = ({ children }: HomeLayoutProps) => {
+  return (
+    <>
+      <TopLogoBar />
+      {children}
+    </>
+  );
+};
+
+export default HomeLayout;
